// libs/list-types/utiac-jr-london-daily-hearing-list/src/models/types.ts
export interface UtiacJrLondonHearing {
  hearingTime: string;
  caseTitle: string;
  representative: string;
  caseReferenceNumber: string;
  judges: string;
  hearingType: string;
  location: string;
  additionalInformation: string;
}

export type UtiacJrLondonHearingList = UtiacJrLondonHearing[];